requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
